Alera entradas e saidas dos metodos para usar long

+++ acervo-biblioteca/src/main/java/br/com/lucalves/acervo_biblioteca/adapters/input/controllers/ILivrosSwagger.java
@@ -9,10 +9,10 @@ import java.util.List;
 public interface ILivrosSwagger {
 
     @PostMapping("adicionar")
-    ResponseEntity<Integer> create(@RequestBody Livro request);
+    ResponseEntity<Long> create(@RequestBody Livro request);
 
     @GetMapping("retorna/{id}")
-    ResponseEntity<Livro> get(@PathVariable("id") Integer id);
+    ResponseEntity<Livro> get(@PathVariable("id") Long id);
 
     @GetMapping("listar-todos")
     ResponseEntity<List<Livro>> getAll();
